:alembic: Experiment with markup-safe notifications
https://github.com/Textualize/textual/issues/5599 seems like a very odd
decision. Yes, of course, `notify` always used markup and anything that
looked like markup would get gobbled up as markup or would confuse the
parser.

But the new parser seems to throw errors if it gets confused. This is... a
weird choice. Imagine if browsers did that. Sure, bad markup can result in a
degraded experience in your average browser but it doesn't crash the
browser, or the "app" you're viewing, most of the time. A parsing engine in
a browser will often try and mitigate the damage.

So... while this doesn't solve the problem for labels and stuff, let's
experiment with a sensible experience degradation approach to unexpected
markup...